refactor : 신고하기 컨텐트 타입에 따라 처리
