+++ src/main/java/com/juu/juulabel/report/ReportType.java
@@ -4,10 +4,11 @@ import lombok.AllArgsConstructor;
 
 @AllArgsConstructor
 public enum ReportType {
-    USER("유저"),
+    MEMBER("멤버"),
     TASTING_NOTE("시음노트"),
+    TASTING_NOTE_COMMENT("시음노트 댓글"),
     DAILY_LIFE("일상생활"),
-    COMMENT("댓글");
+    DAILY_LIFE_COMMENT("일상생활 댓글");
 
     private final String description;
 }
